backdoor/templates/linux/requirements.txt
Pillow==10.2.0
psutil==5.9.8 